Refactor : 상품 상세 리스트 Adapter 분리
